tests: extract storage fixtures from tests/conftest.py
Move fixtures from the monolithic tests/conftest.py into
domain-scoped modules:

- tests/fixtures/storage/storage_classes.py: storage class, OCS,
  HPP, snapshot, volume-mode, RWX-FS, and storage-class-name
  fixtures; RWX_FS_STORAGE_CLASS_NAMES_LIST constant
- tests/fixtures/images/golden_images.py: golden images namespace,
  roles, DVs, data sources, data import crons
- tests/fixtures/storage/data_volumes.py: data volume fixtures

Register new modules in pytest_plugins.

Part of CNV-80951.
